Reorganised project hierarchy, added root folder /manage to handle or product manage logic
