+++ lib/manage/widgets/manage_products_item.dart
